backend/requirements.txt
fastapi==0.123.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.44
aiosqlite==0.20.0
pydantic==2.12.5
pydantic-settings==2.6.1
python-dotenv==1.0.1
sentence-transformers==5.1.2
faiss-cpu==1.13.0


